Add export to csv functionality and extract definition of`LoginData` to separate file
