+++ Cargo.toml
@@ -20,6 +20,7 @@ rpassword = "7.2"
 tokio = { version = "1.30", features = ["time", "rt", "macros", "io-std", "io-util", "rt-multi-thread"] }
 chrono = "0.4"
 prettytable-rs = "0.10"
+csv = "1.2.2"
 
 [profile.release]
 lto = true
