+++ README.md
@@ -5,6 +5,7 @@ Simple password manager written in Rust. Educational purpose. Uses Rust nightly.
 - Passwords are encrypted using `Fernet` with encryption key generated by CSPRNG.
 - Deriving KEK from master password using `Argon2` and wrapping the encryption key with `AES-KW`
 - Automatically exit after short period of inactive using asynchronous programming 
+- Exportng passwords to csv file
 
 ## To do
 ### New features:
@@ -13,7 +14,6 @@ Simple password manager written in Rust. Educational purpose. Uses Rust nightly.
 - Configuration file
 - Command line arguments
 - Password Generation
-- Export passwords to file
 - Change master password
 ### Code:
 - Improve error handling
