+++ src/components/mod.rs
@@ -3,3 +3,5 @@ pub mod cryptography;
 pub mod database;
 pub mod error;
 pub mod metadata;
+pub mod export;
+pub mod logindata;
